+++ components/styles/Dashboard.scss
@@ -5,6 +5,7 @@
   justify-content: space-around;
   align-items: center;
   margin: 2rem;
+  margin-top: 64px;
 }
 
 #button-modal {
