+++ components/styles/Navbar.scss
@@ -26,6 +26,7 @@ a {
 
 .logo {
   font-size: 30px;
+  font-weight: 400;
 }
 
 .settings-icon {
